python/requirements.txt
asgiref==3.5.2
astroid==2.12.10
autopep8==1.7.0
backports.zoneinfo==0.2.1
click==8.1.3
coverage==6.5.0
dill==0.3.5.1
Django==4.1.1
django-environ==0.9.0
django-multiselectfield==0.1.12
djangorestframework==3.14.0
flake8==5.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pylint-plugin-utils==0.7
pytz==2022.2.1
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing-extensions==4.4.0
wrapt==1.14.1
